+++ server/src/main/java/com/server/repository/course/SemesterSubjectRepo.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface SemesterSubjectRepo extends JpaRepository<StudentSubject,Long> {
 
     StudentSubject findByCourseIdAndSemesterNumberAndSubjectId(Long courseId,int semesterNumber,Long subjectId);
-    List<StudentSubject> findAllByCourseIdAndSemesterNumber(Long courseId, int semesterNumber);
+    List<StudentSubject> findAllByCourseIdAndSemesterNumber(Long courseId, Long semesterNumber);
 
 
 }
